discord-bot/requirements.txt
discord.py==2.3.2
fastapi==0.103.0
uvicorn==0.23.2